refactor: delete the proxy and direct-K8s coding-agent dispatch paths
The per-cycle OpenChoreo Component is the only dispatch path now, so the proxy
apply-chain, the direct in-cluster Job dispatcher and the hand-written
Job/ExternalSecret templates have no callers. Collapses the three
coding-dispatch degradation gates into coding-dispatch-oc and locks the boundary
with an arch invariant test.
