+++ services/aep-api/internal/platform/tenant/namespace.go
@@ -51,15 +51,3 @@ func OrgBaseNamespace(orgUUID string) string {
 	salt := hex.EncodeToString(hash[:])[:8]
 	return fmt.Sprintf("wc-%s-%s", strings.ToLower(prefix), salt)
 }
-
-// RemoteWorkerNamespace returns the per-org remote-worker namespace
-// name = OrgBaseNamespace(orgUUID) + "-remote-worker".
-//
-// The `-remote-worker` suffix is the aep fork — wso2cloud's existing
-// shapes all use `-development`, `-staging`, `-production`. The
-// env-less name is intentional: a coding-agent task isn't bound to a
-// user-facing env (per ADR-0001). When wso2cloud-deployement-main
-// pushes a renamed shape we follow.
-func RemoteWorkerNamespace(orgUUID string) string {
-	return OrgBaseNamespace(orgUUID) + "-remote-worker"
-}
